rpc/client: drop neoscan support
It was used as getunspents substitute and it's not longer relevant for NEO 3.
