+++ pkg/rpc/client/client.go
@@ -8,6 +8,7 @@ import (
 	"net"
 	"net/http"
 	"net/url"
+	"sort"
 	"sync"
 	"time"
 
@@ -226,4 +227,35 @@ func (c *Client) Ping() error {
 	}
 	_ = conn.Close()
 	return nil
+}
+
+// unspentsToInputs uses UnspentBalances to create a slice of inputs for a new
+// transcation containing the required amount of asset.
+func unspentsToInputs(utxos state.UnspentBalances, required util.Fixed8) ([]transaction.Input, util.Fixed8, error) {
+	var (
+		num, i   uint16
+		selected = util.Fixed8(0)
+	)
+	sort.Sort(utxos)
+
+	for _, us := range utxos {
+		if selected >= required {
+			break
+		}
+		selected += us.Value
+		num++
+	}
+	if selected < required {
+		return nil, util.Fixed8(0), errors.New("cannot compose inputs for transaction; check sender balance")
+	}
+
+	inputs := make([]transaction.Input, 0, num)
+	for i = 0; i < num; i++ {
+		inputs = append(inputs, transaction.Input{
+			PrevHash:  utxos[i].Tx,
+			PrevIndex: utxos[i].Index,
+		})
+	}
+
+	return inputs, selected, nil
 }
